+++ next-app/src/app/(logged)/dashboard/page.tsx
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import * as React from "react";
+
+export default function DashboardPage() {
+  return <p>To jest twój dashboard!</p>;
+}
