--- next-app/src/app/dashboard/page.tsx
@@ -1,9 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import * as React from "react";
-
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface IChangePasswordPageProps {}
-
-export default function ChangePasswordPage(props: IChangePasswordPageProps) {
-  return <></>;
-}
